refactor: move calendar logic to decorator
